Separate index bindings
The index of the handle currently focused (when a keyboard event is fired) and that of the one about to receive focus (depending on what key was pressed) could be defined using a single binding. However, this may lead to code that is difficult to understand and update so it would be best to have separate bindings for each.